\name{RcmdrPlugin.qual-internal}
\title{Internal RcmdrPlugin.qual objects}
\alias{xbara}
\alias{rchart}
\alias{schart}
\alias{npchart}
\alias{shapeMod}
\alias{pchart}
\alias{shape1}
\alias{paretoMod}
\alias{pareto1}
\alias{xbaro}
\alias{ewMod}
\alias{sum1a}
\alias{sum1Mod}
\alias{hist1Mod}





\description{Internal RcmdrPlugin.qual objects.}
\details{These are not to be called by the user.}
\keyword{internal}
