\name{RcmdrPlugin.qual-package}
\alias{RcmdrPlugin.qual-package}
\alias{RcmdrPlugin.qual}
\docType{package}
\title{
Install the quality  Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for teaching quality control.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.qual\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2009-08-10\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Erin Hodgess <hodgesse@uhd.edu>

Maintainer: Erin Hodgess <hodgesse@uhd.edu>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

